refactor(ChessBoardDto): 패키지 이동
